Refactor auth_client to authenticated_client
- also moved database.ts to utils, still trying to find the right place
