+++ functions/src/google_apis/docs.ts
@@ -1,16 +1,17 @@
 import { docs_v1, google } from 'googleapis';
-import { AuthClient } from '../auth/auth_client';
+import { AuthenticatedClient } from '../auth/authenticated_client';
 
 export class DocsAPI {
   rootFolderId: string = '1poq_tgqfzOF34pJFvdbPgYgI_tD6Mseb';
-  client!: AuthClient;
+  client!: AuthenticatedClient;
   docs!: docs_v1.Docs;
 
-  private constructor() {}
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  private constructor() {};
 
   static async with(uid: string) {
     const docsAPI = new DocsAPI();
-    docsAPI.client = await AuthClient.getInstanceWithCredentialsFor(uid);
+    docsAPI.client = await AuthenticatedClient.getInstanceFor(uid);
     docsAPI.docs = google.docs({version: 'v1', auth: docsAPI.client.getOAuth2Client()});
   }
 
@@ -18,7 +19,7 @@ export class DocsAPI {
     const createResponse = await this.docs.documents.create({
       requestBody: {
         'title': title,
-      }
+      },
     });
 
     return createResponse.data;
